backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
openai
email-validator
bcrypt
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
